extracted configuration and added firt bytes download
