+++ condow_core/src/machinery/tests.rs
@@ -1,9 +1,9 @@
-mod download {
+mod download_chunks {
     use crate::{
         condow_client::{failing_client_simulator::FailingClientSimulatorBuilder, IgnoreLocation},
         config::Config,
         errors::CondowErrorKind,
-        machinery::download_range,
+        machinery::download_chunks,
     };
 
     #[tokio::test]
@@ -22,7 +22,8 @@ mod download {
             .condow(config)
             .unwrap();
 
-        let result = download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+        let result =
+            download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
         let stream = result.unwrap();
 
@@ -49,7 +50,8 @@ mod download {
             .condow(config)
             .unwrap();
 
-        let result = download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+        let result =
+            download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
         let stream = result.unwrap();
 
@@ -77,7 +79,8 @@ mod download {
             .condow(config)
             .unwrap();
 
-        let result = download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+        let result =
+            download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
         let stream = result.unwrap();
 
@@ -108,7 +111,7 @@ mod download {
                 .unwrap();
 
             let result =
-                download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+                download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
             let stream = result.unwrap();
 
@@ -143,7 +146,7 @@ mod download {
                 .unwrap();
 
             let result =
-                download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+                download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
             let stream = result.unwrap();
 
@@ -176,7 +179,7 @@ mod download {
                 .unwrap();
 
             let result =
-                download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+                download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
             let stream = result.unwrap();
 
@@ -211,7 +214,7 @@ mod download {
                 .unwrap();
 
             let result =
-                download_range(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
+                download_chunks(condow.client, condow.config, IgnoreLocation, 0..100, ()).await;
 
             let stream = result.unwrap();
 
@@ -222,18 +225,8 @@ mod download {
             //assert_eq!(err.msg(), "panicked while retrying");
         }
     }
-}
 
-mod download_chunks {
-    use tracing::Span;
-
-    use crate::{
-        condow_client::IgnoreLocation,
-        config::Config,
-        machinery::{download_chunks, part_request::PartRequestIterator, DownloadSpanGuard},
-        test_utils::*,
-        InclusiveRange,
-    };
+    /*
 
     #[tokio::test]
     async fn from_0_to_inclusive_range_smaller_than_part_size() {
@@ -321,4 +314,5 @@ mod download_chunks {
 
         assert_eq!(&result, &data[range.to_std_range_usize()]);
     }
+    */
 }
